requirements/base.txt
# SHA1:a9dde048f1ee1f00586264d726d0e89f16e56183
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-e file:.
    # via -r requirements/base.in
accept-types==0.4.1
    # via dj
amqp==5.1.1
    # via kombu
anyio==3.6.1
    # via
    #   starlette
    #   watchfiles
asciidag==0.2.0
    # via dj
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
cachelib==0.9.0
    # via dj
celery==5.2.7
    # via dj
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
commonmark==0.9.1
    # via rich
deprecated==1.2.13
    # via redis
docopt==0.6.2
    # via dj
fastapi==0.70.1
    # via dj
graphql-core==3.2.3
    # via strawberry-graphql
greenlet==1.1.3
    # via sqlalchemy
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
kombu==5.2.4
    # via celery
msgpack==1.0.4
    # via dj
multidict==6.0.2
    # via yarl
packaging==21.3
    # via redis
prompt-toolkit==3.0.31
    # via click-repl
pydantic==1.10.2
    # via
    #   fastapi
    #   sqlmodel
pygments==2.13.0
    # via rich
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via strawberry-graphql
python-dotenv==0.19.2
    # via dj
pytz==2022.4
    # via celery
pyyaml==6.0
    # via dj
redis==4.3.4
    # via dj
requests==2.28.1
    # via dj
rich==12.6.0
    # via dj
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sniffio==1.3.0
    # via anyio
sqlalchemy==1.4.35
    # via
    #   dj
    #   sqlalchemy-utils
    #   sqlmodel
sqlalchemy-utils==0.38.3
    # via dj
sqlalchemy2-stubs==0.0.2a27
    # via sqlmodel
sqlmodel==0.0.5
    # via dj
sqloxide==0.1.15
    # via dj
sqlparse==0.4.3
    # via dj
starlette==0.16.0
    # via fastapi
strawberry-graphql==0.133.5
    # via dj
typing-extensions==4.4.0
    # via
    #   pydantic
    #   sqlalchemy2-stubs
    #   strawberry-graphql
urllib3==1.26.12
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.12
    # via dj
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.14.1
    # via deprecated
yarl==1.8.1
    # via dj
